+++ manifest.json
@@ -23,17 +23,13 @@
   },
   "content_scripts": [
     {
-      "matches": ["https://*.instructure.com/courses/*"],
-      "js": ["scripts/associateSectionsToBlueprints/associatedBpCourse.js"],
-      "type": "module"
-    },
-    {
       "matches": [
-        "https://byui.instructure.com/*"
+        "https://*.instructure.com/courses/*"
       ],
       "js": [
-        "./scripts/copyAllBrokenLinks/contentScript.js"
-      ]
+        "scripts/associateSectionsToBlueprints/associatedBpCourse.js"
+      ],
+      "type": "module"
     }
   ]
 }
