+++ scripts/copyAllBrokenLinks/background.js
