+++ scripts/copyAllBrokenLinks/copyAllBrokenLinksPopup.html
@@ -2,7 +2,7 @@
 <html>
 <head>
   <link rel="stylesheet" type="text/css" href="copyAllBrokenLinksPopup.css">
-  <script src="copyAllBrokenLinks.js"></script>
+  <script src="copyAllBrokenLinksPopup.js"></script>
 </head>
 <title>Copy All Broken Links</title>
 <body>
