+++ scripts/copyAllBrokenLinks/copyAllBrokenLinksPopup.js
